Move testFit tests
